requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.29.0
shimmy>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# SUMO Python bindings (included with SUMO installation)
# traci - automatically available when SUMO_HOME is set

# HTTP/API calls (for OSM download)
requests>=2.31.0
urllib3>=2.0.0

# Geospatial
osmium>=3.6.0
pyproj>=3.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
tabulate>=0.9.0

# Logging
loguru>=0.7.0
